% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{agegrp_name}
\alias{agegrp_name}
\title{Generate names for age-group bands}
\usage{
agegrp_name(
  min_age = 0L,
  max_age = 85L,
  grp_width = 5L,
  grp_lessthan_1 = TRUE,
  match_input = FALSE,
  match_input_max_age = max_age
)
}
\arguments{
\item{min_age}{A non-negative integer. The lower age limit for which
names will be generated.}

\item{max_age}{A non-negative integer. The upper age limit for which
names will be generated.}

\item{grp_width}{A positive integer. The band width of the age-groups.}

\item{grp_lessthan_1}{A logical scalar. if \code{TRUE} and
\code{min_age == 0}, then the first age-group name is "<1".}

\item{match_input}{A logical scalar. If \code{TRUE}, then the names
are repeated to match every single year of age between \code{min_age}
and \code{match_input_max_age}.}

\item{match_input_max_age}{a non-negative integer. See above.}
}
\value{
A character vector of with the names for the age-groups.
}
\description{
`agegrp_name` generates names for age-group bands given lower and upper
  age limits, and band width
}
\examples{
agegrp_name(20, 79, 5)
agegrp_name(20, 80, 5)
agegrp_name(0, 80, 10, TRUE)
agegrp_name(20, 30, 5, FALSE, TRUE)
agegrp_name(20, 30, 5, FALSE, TRUE)
agegrp_name(20, 30, 5, FALSE, TRUE, 32)
}
