% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clone_dt}
\alias{clone_dt}
\title{Clone a data.table}
\usage{
clone_dt(dt, times, idcol = TRUE)
}
\description{
`clone_dt` clones a data.table and binds the copies at the bottom of
the original data.table. It also creates an column named \code{`.id`}
to identify each iteration. The keys of the input data.table is retained.
}
