% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crossval_gamlss}
\alias{crossval_gamlss}
\title{Deterministic prediction from a gamlss object}
\usage{
crossval_gamlss(dt, gamlss_obj, orig_data = dt, colnam = "rank")
}
\description{
`crossval_gamlss` returns the observed and predicted values of the dependent
 variable. Useful for cross-validation metrics.
}
