% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dependencies}
\alias{dependencies}
\title{Simplified loading and installing of packages}
\usage{
dependencies(
  pkges,
  install = TRUE,
  update = FALSE,
  quiet = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{pkges}{a character vector with the names of the packages to load.}

\item{install}{if TRUE (default), any packages not already installed will be.}

\item{update}{if TRUE, this function will install a newer version of the
package if available.}

\item{quiet}{if TRUE (default), package startup messages will be suppressed.}

\item{verbose}{if TRUE (default), diagnostic messages will be printed.}

\item{...}{other parameters passed to \code{\link{require}},
\code{\link{install.packages}}, and
\code{\link{available.packages}}.}
}
\value{
a data frame with four columns and rownames corresponding to the
        packages to be loaded. The four columns are: loaded (logical
        indicating whether the package was successfully loaded), installed
        (logical indicating that the package was installed or updated),
        loaded.version (the version string of the installed package), and
        available.version (the version string of the package currently
        available on CRAN). Note that this only reflects packages listed in
        the \code{pkges} parameter. Other packages may be loaded and/or
        installed as necessary by \code{install.packages} and \code{require}.
        If \code{verbose=FALSE} the data frame will be returned using
        \code{\link{invisible}}.
}
\description{
This is a wrapper to \code{\link{require}} and \code{\link{install.packages}}.
Specifically, this will first try to load the package(s) and if not found
it will install then load and attach the packages. Additionally, if the
\code{update=TRUE} parameter is specified it will check the currently
installed package version with what is available on CRAN (or mirror) and
install the newer version.
}
\details{
The function was originally created by Jason Bryer
\href{https://www.r-bloggers.com/function-to-simplify-loading-and-installing-packages/}{here}
and the source is available \href{https://gist.github.com/jbryer/9112634}{here}.
Note: I renamed the function to \code{dependencies} and adapted it to attach
instead of only load the packages.
}
\examples{
\dontrun{
dependencies(c('devtools','lattice','ggplot2','psych'))
}
}
