% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_dropbox_path}
\alias{get_dropbox_path}
\title{Get Dropbox path}
\usage{
get_dropbox_path(pathtail = character(0), type = c("personal", "business"))
}
\arguments{
\item{pathtail}{A String vector (if not a string then it is converted to String).
If present, it gets concatenated with the Dropbox path.
See examples.}

\item{type}{A String scalar ("personal" or "business"). Which Dropbox path to return? The personal or the business one? It may be abbreviated.}
}
\value{
Dropbax path as a String. If pathtail is present, it concatenates the Dropbox path with pathtail.
}
\description{
`get_dropbox_path` returns the path of Dropbox. Works for both personal and business accounts
}
\details{
This is an auxilliary function: It finds the Dropbox path in Windows, Linux, and OSX operating systems.
}
\examples{
\dontrun{
# Only work if Dropbox is installed.
get_dropbox_path() # Returns personal Dropbox path
get_dropbox_path(type = "business") # Returns business Dropbox path
get_dropbox_path("pathdownthetree") # Returns "Dropbox_path/pathdownthetree",
 # where Dropbox_path is the path to personal Dropbox
}
}
