% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pcloud_path}
\alias{get_pcloud_path}
\title{Get pCloud path}
\usage{
get_pcloud_path(pathtail = character(0))
}
\arguments{
\item{pathtail}{A String vector (if not a string then it is converted to String).
If present, it gets concatenated with the pCloud path.
See examples.}
}
\value{
pCloud path as a String. If pathtail is present, it concatenates the Dropbox path with pathtail.
}
\description{
`get_pcloud_path` returns the path of pCloud
}
\details{
This is an auxilliary function: It finds the pCloud path in Windows, Linux, and OSX operating systems.
}
\examples{
\dontrun{
# Only work if Dropbox is installed.
get_pcloud_path() # Returns pCloud path
get_pcloud_path("pathdownthetree") # Returns "pcloud_path_path/pathdownthetree",
# where pcloud_path is the path to pCloud
}
}
