% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{guess_gamlss}
\alias{guess_gamlss}
\title{Prediction from a gamlss object in parallel}
\usage{
guess_gamlss(dt, gamlss_obj, orig_data = gamlss_obj$data, nc = 1L)
}
\description{
`guess_gamlss` returns a data.table with the predicted
 variable. `dt` needs to have a column with percentiles named `rank_y`,
 where `y` the name of the predicted variable (i.e. bmi).
}
