% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pctl_rank}
\alias{pctl_rank}
\title{Calculate percentile rank}
\usage{
pctl_rank(
  x,
  ties.method = c("average", "first", "random", "max", "min", "dense")
)
}
\description{
`pctl_rank` calculates the percentile rank of a numeric vector
}
