% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reldist_diagnostics}
\alias{reldist_diagnostics}
\title{Compare two distributions}
\usage{
reldist_diagnostics(
  comparison,
  reference,
  comparison_wt,
  reference_wt,
  main,
  smooth = 0.35,
  discrete = FALSE
)
}
\description{
Summary statistics for the location/shape decomposition of the relative
distribution of the exposure: Modelled to Observed."
}
